fix(lint): unbreak the 2000-line file-size gate on main
Two files crossed the cap in the 2026-08-11 batch and `lint` has been red
on `main` ever since:

  crates/perry-hir/src/lower/pre_scan.rs   1985 -> 2011  (#7828)
  crates/perry-runtime/src/gc/layout.rs    1982 -> 2023  (#7809, #7812)

Both are pure code moves, no logic change:

* `pre_scan_weakref_locals` and its doc comment move to
  `lower/pre_scan/weakref_locals.rs` (1653 lines left behind). It is the
  cohesive unit — one top-level pre-scan with its own five local sets —
  and it is where #7828 added the lines.

* `LayoutSlotMask` (the enum and its whole `impl`) moves to
  `gc/layout/slot_mask.rs` (1807 lines left behind). Its visibility widens
  from `pub(super)` to `pub(in crate::gc)` because the type is now one
  module deeper and `layout_tables.rs` / `hot_tls.rs` still name it; the
  reachable set is unchanged.

`scripts/check_file_size.sh` passes.
